Framing and support for the echo command
Added framing for the Client TcpStream to allow for parsing and writing
messages to the TcpStream

Also, added simple separation of duties between the "server" listening
and handling each client connection. Not sure if I love the design thus
far, but it will work in the meantime. There may need to be a way for
the server to keep track of clients or be able to control their
connections, but we can cross that bridge if needed. There will also
need to be some sort of state eventually that allows for the setting /
getting of keys from the database.

Finally, added support for the ECHO command to solve this level of
codecrafters