Add support for GET / SET command
Adds the actual DB for getting and setting values in the DB.
Design decision was to pass around an Arc<RwLock<HashMap<String,
String>>> to provide each client serialized access to the DB, but that
very well may need to change in the future.

One way to change it is have a main server 'loop' that owns the actual
key/value map that can then read from a channel for requests to get/set
keys. That way, in the future pub/sub model, the main server loop will
be aware of changes and can notify any subscribers. For now, this works
out.