Add support for expiring keys in the db
Created a modified DB value to allow us to communicate whether a key
expires (or not) and when. We make the decision on subsequent "GET"s

Main issue: we never remove those keys from the DB. Future work needs to
be done there. Options: we can get a write handle to the DB when we
identify an expired key and expire it on the GET (kind of wonky in my
opinion; would slow down our GET operations if we then have to wait for
a write lock on the DB), we setup a separate task that will periodically
clear out the database with keys that are expired, or we move to the
other design thought in the GET / SET commit where there is a separate
task that simply manages the DB access and will have ownership over the
DB during any GET operations and can make those clears as needed.