Major refactor before moving on
After completing the main challenge, I leveraged some LLMs to do some
code analysis and got some improvements on how to improve the previous
commit. First main point was to stop the RespValue usage of String to
avoid needing extra allocations. That basically means that Nom was out
if I was going to use it the way I wanted to. Nom didn't seem to play
nicely with the Bytes type. I searched around the codecrafters forums
and found the blog here: https://dpbriggs.ca/blog/Implementing-A-Copyless-Redis-Protocol-in-Rust-With-Parsing-Combinators/
which basically did exactly what I wanted to do. The parser is heavily
inspired by that blog (almost to the point of copying it) with some
minor tweaks. Nom probably could've worked, it just would have required
two full passes through the Nom combinators and I didn't like that much.
Nom seems to play better with fully owned values and outputting
something that references &[u8] didn't seem like the best way to do
things. I likely could have created a Bytes type from that but I bet
there would have still been some copying. This approach uses the buffer
we receive from the network.

Other enhancements include using DashMap over RwLock<HashMap<..>> and
transforming the command parsing into separate types to handle the
commands. I also included a key expiration task that would watch the DB
and be notified of new keys to expire to ensure that we actually expire
keys when they hit their expiration. We still have to check at the time
of a GET command, but this will ensure that if someone SETs a bunch of
keys and then never calls GET on them, we don't just have a full map.