Add basic support for RPUSH command
Allows for lists to be stored in the Redis database. To support this,
the database type needed to be extracted out and allowed for multiple
DashMaps to be contained within the "Database". Currently, we have an
API that controls the access to the various DashMaps to hopefully reduce
the risk of deadlock by holding on to references, but we will see if
that actually happens. DashMap is great but it may require some other
locking behavior / design if deadlocks occur with two simultaneous
clients.